refactor: 전체적인 코드 정리(줄바꿈, 주석삭제 등)
