+++ src/components/Layout.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+
 import styled from 'styled-components'
 import { colors } from '@/styles/colors'
