+++ src/router/index.tsx
@@ -1,9 +1,10 @@
-import React from 'react';
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import { Main } from '@/page/main';
-import { SignIn, SignUp } from '@/page/sign';
-import { Todos } from '@/page/todos';
-import PrivateRoute from '@/components/PrivateRoute';
+import React from 'react'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
+
+import { Main } from '@/page/main'
+import { SignIn, SignUp } from '@/page/sign'
+import { Todos } from '@/page/todos'
+import PrivateRoute from '@/components/PrivateRoute'
 
 export const Router = () => {
   return (
