+++ src/styles/font/index.ts
@@ -1,6 +1,5 @@
 import { css } from 'styled-components'
-
-import { fontWeight } from '@/typography'
+import { fontWeight } from '@/styles/typography'
 
 export const FONT_FAMILY_NAME = 'Spoqa Han Sans Neo'
