+++ src/styles/typography/index.ts
@@ -1,7 +1,9 @@
 export type { FontSizeKey } from './font-size';
 export { fontSize, fontSizeCSS } from './font-size';
+
 export type { FontWeightKey } from './font-weight';
 export { fontWeight, fontWeightCSS } from './font-weight';
+
 export { getLineClampCSS } from './get-line-clamp-css';
 export type { TextDecorationKey } from './text-decoration-css';
 export { textDecoration, textDecorationCSS } from './text-decoration-css';
